requirements.txt
torch
matplotlib
pyyaml